backend-recognition/requirements.txt
fastapi
uvicorn
pydantic
insightface
numpy
faiss-cpu
opencv-python
pillow
python-multipart
onnxruntime